requirements.txt
# cffi==1.14.3
# chardet==3.0.4
# click==7.1.2
# coloredlogs==14.0
# cryptography==3.1.1
# fdf==0.1.9
Flask==1.1.2
fpdf==1.7.2
# humanfriendly==8.2
# img2pdf==0.4.0
# itsdangerous==1.1.0
# Jinja2==2.11.2
# lxml==4.6.1
# MarkupSafe==1.1.1
ocrmypdf==11.2.1
# pdfminer.six==20200726
# pikepdf==1.19.3
Pillow==8.0.0
# pluggy==0.13.1
# pycparser==2.20
PyPDF2==1.26.0
pytesseract==0.3.6
# reportlab==3.5.53
# six==1.15.0
# sortedcontainers==2.2.2
# tqdm==4.50.2
# Werkzeug==1.0.1
